alteracoes de migracao

+++ db/migrate/20240916053913_remove_credit_and_debit_from_users.rb
@@ -1,6 +1,9 @@
 class RemoveCreditAndDebitFromUsers < ActiveRecord::Migration[6.1]
   def change
-    remove_column :users, :credit, :decimal, precision: 10, scale: 2, default: "0.0", null: false
-    remove_column :users, :debit, :decimal, precision: 10, scale: 2, default: "0.0", null: false
+    # Remove a coluna credit se ela existir
+    remove_column :users, :credit, :decimal, precision: 10, scale: 2, default: "0.0", null: false if column_exists?(:users, :credit)
+
+    # Remove a coluna debit se ela existir
+    remove_column :users, :debit, :decimal, precision: 10, scale: 2, default: "0.0", null: false if column_exists?(:users, :debit)
   end
 end
